[feat]: integrate with redis to store statistic results

+++ Makefile
@@ -2,4 +2,4 @@ run:
 	go run main.go
 
 clean:
-	rm -rf repos
+	clear && rm -rf repos
